+++ src/neospy/__init__.py
@@ -1,15 +1,20 @@
-from .vector import Vector, Frames, State, CometElements
+from .vector import (
+    Vector,
+    Frames,
+    State,
+    CometElements,
+    SimultaneousStates,
+)
 from . import (
     constants,
     covariance,
     wise,
     neos,
-    data,
+    cache,
     population,
     flux,
     mpc,
     irsa,
-    pds,
     ztf,
     fov,
 )
@@ -31,18 +36,14 @@ from .conversion import (
 )
 from .horizons import HorizonsProperties
 
-
-# pylint: disable-next=import-error
-from ._core import SimultaneousStates  # type: ignore
-
 import logging
 
 
 __all__ = [
+    "cache",
     "constants",
     "CometElements",
     "covariance",
-    "data",
     "irsa",
     "Frames",
     "moid",
@@ -57,7 +58,6 @@ __all__ = [
     "wise",
     "neos",
     "mpc",
-    "pds",
     "SimultaneousStates",
     "propagate_n_body",
     "propagate_two_body",