+++ src/neospy_core/src/fov/neos.rs
@@ -1,7 +1,7 @@
 //! # NEOS field of views
 use super::{closest_inside, Contains, FovLike, OnSkyRectangle, SkyPatch, FOV};
 use crate::constants::{NEOS_HEIGHT, NEOS_WIDTH};
-use crate::frames::rotate_around;
+// use crate::frames::rotate_around;
 use crate::prelude::*;
 use nalgebra::Vector3;
 use serde::{Deserialize, Serialize};
@@ -100,6 +100,15 @@ impl FovLike for NeosCmos {
     fn n_patches(&self) -> usize {
         1
     }
+
+    fn try_frame_change_mut(
+        &mut self,
+        target_frame: Frame,
+    ) -> Result<(), NEOSpyError> {
+        self.observer.try_change_frame_mut(target_frame)?;
+        self.patch = self.patch.try_frame_change(target_frame)?;
+        Ok(())
+    }
 }
 
 /// NEOS frame data, all 4 chips of a visit.
@@ -183,61 +192,61 @@ impl NeosVisit {
         })
     }
 
-    /// x_width is the longer dimension in radians
-    pub fn from_pointing(
-        x_width: f64,
-        y_width: f64,
-        gap_fraction: f64,
-        pointing: Vector3<f64>,
-        rotation: f64,
-        observer: State,
-        side_id: u16,
-        stack_id: u8,
-        quad_id: u8,
-        loop_id: u8,
-        subloop_id: u8,
-        exposure_id: u8,
-        cmos_id: u8,
-        band: u8,
-    ) -> Self {
-        let chip_x_width = (1.0 - 3.0 * gap_fraction) * x_width / 4.0;
+    // /// x_width is the longer dimension in radians
+    // pub fn from_pointing(
+    //     x_width: f64,
+    //     y_width: f64,
+    //     gap_fraction: f64,
+    //     pointing: Vector3<f64>,
+    //     rotation: f64,
+    //     observer: State,
+    //     side_id: u16,
+    //     stack_id: u8,
+    //     quad_id: u8,
+    //     loop_id: u8,
+    //     subloop_id: u8,
+    //     exposure_id: u8,
+    //     cmos_id: u8,
+    //     band: u8,
+    // ) -> Self {
+    //     let chip_x_width = (1.0 - 3.0 * gap_fraction) * x_width / 4.0;
 
-        // Rotate the Z axis to match the defined rotation angle, this vector is not
-        // orthogonal to the pointing vector, but is in the correct plane of the final
-        // up vector.
-        let up_vec = rotate_around(&Vector3::new(0.0, 0.0, 1.0), pointing, -rotation);
+    //     // Rotate the Z axis to match the defined rotation angle, this vector is not
+    //     // orthogonal to the pointing vector, but is in the correct plane of the final
+    //     // up vector.
+    //     let up_vec = rotate_around(&Vector3::new(0.0, 0.0, 1.0), pointing, -rotation);
 
-        // construct the vector orthogonal to the pointing and rotate z axis vectors.
-        // left = cross(up, pointing)
-        let left_vec = pointing.cross(&up_vec);
+    //     // construct the vector orthogonal to the pointing and rotate z axis vectors.
+    //     // left = cross(up, pointing)
+    //     let left_vec = pointing.cross(&up_vec);
 
-        // Given the new left vector, and the existing orthogonal pointing vector,
-        // construct a new up vector which is in the same plane as it was before, but now
-        // orthogonal to the two existing vectors.
-        // up = cross(pointing, left)
-        let up_vec = pointing.cross(&left_vec);
+    //     // Given the new left vector, and the existing orthogonal pointing vector,
+    //     // construct a new up vector which is in the same plane as it was before, but now
+    //     // orthogonal to the two existing vectors.
+    //     // up = cross(pointing, left)
+    //     let up_vec = pointing.cross(&left_vec);
 
-        // +------+-+------+-+------+-+------+   ^
-        // |  1   |g|  2   |g|  3   |g|  4   |   |
-        // |      |a|      |a|      |a|      |   y
-        // |      |p|      |p|      |p|      |   |
-        // +------+-+------+-+------+-+------+   _
-        // <-cf->    x ->
-        //
-        // pointing vector is in the middle of the 'a' in the central gap.
+    //     // +------+-+------+-+------+-+------+   ^
+    //     // |  1   |g|  2   |g|  3   |g|  4   |   |
+    //     // |      |a|      |a|      |a|      |   y
+    //     // |      |p|      |p|      |p|      |   |
+    //     // +------+-+------+-+------+-+------+   _
+    //     // <-cf->    x ->
+    //     //
+    //     // pointing vector is in the middle of the 'a' in the central gap.
 
-        let outer: Vector3<f64> = rotate_around(&left_vec, up_vec, -lon_width / 2.0);
-        let n2: Vector3<f64> = rotate_around(&(-left_vec), up_vec, lon_width / 2.0);
+    //     let outer: Vector3<f64> = rotate_around(&left_vec, up_vec, -lon_width / 2.0);
+    //     let n2: Vector3<f64> = rotate_around(&(-left_vec), up_vec, lon_width / 2.0);
 
-        let long_top: Vector3<f64> = rotate_around(&up_vec, left_vec, y_width / 2.0);
-        let long_bottom: Vector3<f64> = rotate_around(&(-up_vec), left_vec, -y_width / 2.0);
+    //     let long_top: Vector3<f64> = rotate_around(&up_vec, left_vec, y_width / 2.0);
+    //     let long_bottom: Vector3<f64> = rotate_around(&(-up_vec), left_vec, -y_width / 2.0);
 
-        // construct the 4 normal vectors
-        Self {
-            edge_normals: [n1.into(), n2.into(), n3.into(), n4.into()],
-            frame,
-        }
-    }
+    //     // construct the 4 normal vectors
+    //     Self {
+    //         edge_normals: [n1.into(), n2.into(), n3.into(), n4.into()],
+    //         frame,
+    //     }
+    // }
 }
 
 impl FovLike for NeosVisit {
@@ -260,6 +269,16 @@ impl FovLike for NeosVisit {
     }
 
     fn n_patches(&self) -> usize {
-        8
+        4
+    }
+
+    fn try_frame_change_mut(&mut self, target_frame: Frame) -> Result<(), NEOSpyError> {
+        let _ = self
+            .chips
+            .iter_mut()
+            .map(|ccd| ccd.try_frame_change_mut(target_frame))
+            .collect::<Result<Vec<_>, _>>()?;
+        self.observer.try_change_frame_mut(target_frame)?;
+        Ok(())
     }
 }