+++ src/neospy_core/src/propagation/mod.rs
@@ -75,7 +75,7 @@ pub fn propagate_n_body_spk(
     let frame = state.frame;
     let spk = get_spk_singleton().try_read().unwrap();
     spk.try_change_center(&mut state, 0)?;
-    state.try_change_frame(Frame::Ecliptic)?;
+    state.try_change_frame_mut(Frame::Ecliptic)?;
 
     let mass_list = {
         if include_extended {
@@ -111,7 +111,7 @@ pub fn propagate_n_body_spk(
         0,
     );
     spk.try_change_center(&mut new_state, center)?;
-    new_state.try_change_frame(frame)?;
+    new_state.try_change_frame_mut(frame)?;
     Ok(new_state)
 }
 
@@ -163,7 +163,7 @@ pub fn propagation_n_body_vec(
 
     for mut state in states.into_iter() {
         spk.try_change_center(&mut state, 0)?;
-        state.try_change_frame(Frame::Ecliptic)?;
+        state.try_change_frame_mut(Frame::Ecliptic)?;
         if jd_init != state.jd {
             return Err(NEOSpyError::ValueError(
                 "All input states must have the same JD".into(),